RESTORE: Move all backend files back to backend/ directory
- Move all Python files from root back to backend/ directory
- Restore all imports to 'from backend.module import' format
- Update CI workflow to reference backend/ directory structure
- Move .flake8 config back to backend/
- Restore original project structure per user request
